assistify/backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==4.2.10
django-cors-headers==4.5.0
django-crispy-forms==2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.8
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mysql-connector-python==8.3.0
packaging==24.1
pillow==11.0.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
faker
markdown
django-filter
gunicorn==21.2.0
psycopg2-binary==2.9.9
openai